닐슨 쇼핑 다이어리 서치 관리자 구현 완료

+++ src/main/java/com/enuri/nielsen/admin/domain/shoppingDiary/compositePrimaryKeyClass/BuyHistoryId.java
@@ -18,5 +18,5 @@ public class BuyHistoryId implements Serializable {
     private long integrationBuyNo;
 
     @Column(name = "SRC_TBL_ID", nullable = false)
-    private int sourceTableId;
+    private long sourceTableId;
 }
